Fix behaviour when fmt param is null.
When the fmt param is null, the code assumes that the first args element is the new fmt.

This causes a problem when the arg[0] is also null, throwing an Error.

This behaviour is inconsistent with format(null, ['foo', null]) that doesnt throw and returns 'foo null'.

This commit tries to fix that. A call like format(null, [null, 'foo']) will return 'null foo'.